Add support for NOLINTBEGIN/END
Add support for specifying a NOLINT in a block of code. The
block is started with a NOLINTBEGIN comment which can specify
one or more categories. The block is ended with a NOLINTEND
comment which clears all categories specified in the BEGIN.

Errors are generated if a BEGIN statment occurs in a block, if
END is specified without a BEGIN or if END specifies categories

This is similar to the clang-tidy NOLINTBEGIN/END feature.